api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
openai==1.52.2
httpx==0.27.2
pillow==10.4.0